Fix msgId type

+++ mcrouter/lib/network/AsyncMcClientImpl.h
@@ -166,7 +166,7 @@ class AsyncMcClientImpl :
 
   void sendFakeReply(McClientRequestContextBase& request);
 
-  static void incMsgId(size_t& msgId);
+  static void incMsgId(uint64_t& msgId);
 
   // Returns the next request id to be used for cycles measurement.
   static uint64_t nextCyclesRequestId();
